requirements.txt
pip==21.1.2

wheel==0.41.3
apache-airflow[amazon]==2.7.2
openpyxl==3.1.2
pydevd-pycharm~=232.10072.31
pandas
apache-airflow-providers-snowflake==5.1.0